+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 **Breaking**, **New**, **Fix**, **Change**, **Others**
 
+## 1.0.0-alpha16 (2025-11-14)
+- **New** [POSWEB-1503](https://posguru.atlassian.net/browse/POSWEB-1503) add latitude and longitude
+  - added  latitude and longitude to partner and location
+
 ## 1.0.0-alpha15 (2025-11-10)
 - **Fix** [POSWEB-1451](https://posguru.atlassian.net/browse/POSWEB-1451) apply service charge to order
   - apply delivery charge